requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
APScheduler==3.10.4
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.9.1
blinker==1.6.2
Brotli==1.1.0
CacheControl==0.13.1
cachelib==0.9.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudinary==1.36.0
configparser==6.0.0
contourpy==1.1.1
cryptography==41.0.4
cycler==0.12.0
decorator==5.1.1
distro==1.8.0
executing==2.0.0
fake-useragent==1.3.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-Compress==1.14
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.43.0
frozenlist==1.4.0
fuzzywuzzy==0.18.0
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.1
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
httpcore==1.0.1
httplib2==0.22.0
httpx==0.25.1
idna==3.4
ipython==8.16.1
iso8601==2.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonify==0.5
kiwisolver==1.4.5
Levenshtein==0.21.1
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
msgpack==1.0.6
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
mysqlclient==2.2.0
numpy==1.25.2
openai==1.14.1
outcome==1.3.0
packaging==23.1
pandas==2.1.0
paramiko==3.3.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pipdeptree==2.13.0
platformdirs==3.10.0
plotly==5.16.1
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.8.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.21.1
pytimeparse2==1.7.1
pytz==2023.3
rapidfuzz==3.3.0
redis==5.0.0
requests==2.31.0
row==0.1.0.dev3
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
selenium==4.14.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spotipy==2.23.0
SQLAlchemy==2.0.22
sqlite3-to-mysql==2.1.0
sshtunnel==0.4.0
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
traitlets==5.11.2
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==2.0.5
wcwidth==0.2.8
webdriver-manager==4.0.1
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.2
